Rework pairing server, so that each implementation does not need to know which messages get sent/received - they just implement a Socket interface.
This required extensive refactoring of the tests for pairing server, but should increase the fidelity of these tests as we actually make assertions on sent/received messages.

The pairing server API is also simpler, being a single async function for each of create/respond. The testing becomes clean when firing off multiple async functions and awaiting the results.